DAGs/Airflow_V2_tested_PY_3_8/requirements.txt
asyncio
requests
google-cloud-pubsub
aiohttp >= 3.7.4.post0
google_cloud_storage
pandas